+++ src/common/constants.py
@@ -132,6 +132,7 @@ TYPE_EXPORT_METADATA = "Export Metadata"
 TYPE_COMPLETE_SUB = "Complete Submission"
 TYPE_CROSS_SUBMISSION = "Validate Cross-submission"
 TYPE_GENERATE_DCF = "Generate DCF Manifest"
+RESTORE_DELETED_DATA_FILE = "Restore Deleted Data File"
 FAILED = "Failed"
 ADDITION_ERRORS = "additionalErrors"
 STUDY_ABBREVIATION = "studyAbbreviation"
